# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ESTADIA API — detention time tracking and billing for truck drivers
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /perfil:
    get:
      operationId: getPerfil
      tags: [profile]
      summary: Get current driver profile
      responses:
        "200":
          description: Driver profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Motorista"
        "401":
          description: Not authenticated
    put:
      operationId: updatePerfil
      tags: [profile]
      summary: Update driver profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MotoristaUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Motorista"
...
  /esperas:
    get:
      operationId: listEsperas
      tags: [esperas]
      summary: List all waiting records for current driver
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [aguardando, encerrada, cobranca_gerada]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of esperas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EsperaList"
    post:
      operationId: createEspera
      tags: [esperas]
      summary: Create a new waiting record (records server timestamp)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EsperaInput"
      responses:
        "201":
          description: Created espera
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Espera"
  /esperas/{id}:
    get:
      operationId: getEspera
      tags: [esperas]
      summary: Get a specific waiting record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Espera details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Espera"
  /esperas/{id}/encerrar:
    patch:
      operationId: encerrarEspera
      tags: [esperas]
      summary: Close a waiting record (record departure time)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EncerrarEsperaInput"
      responses:
        "200":
          description: Closed espera
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Espera"
  /esperas/{id}/fotos:
    post:
      operationId: uploadFoto
      tags: [esperas]
      summary: Upload a photo for a waiting record (base64 encoded)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FotoInput"
      responses:
        "200":
          description: Photo uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FotoResult"
...
  /public/verificar/{token}:
    get:
      operationId: verificarCobranca
      tags: [public]
      summary: Public verification page — returns immutable record data
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Verified billing record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificacaoPublica"
        "404":
          description: Not found

  # ─── SUBSCRIPTIONS ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    Motorista:
      type: object
      required: [id, telefone, plano, created_at]
      properties:
        id:
          type: string
        telefone:
          type: string
        nome:
          type: ["string", "null"]
        tipo:
          type: ["string", "null"]
          enum: ["TAC autônomo", "ETC frota", null]
        plano:
          type: string
          enum: [gratis, pro_mensal, pro_anual]
        created_at:
          type: string
          format: date-time
    MotoristaUpdate:
      type: object
      properties:
        nome:
          type: string
        tipo:
          type: string
          enum: ["TAC autônomo", "ETC frota"]
...
    Veiculo:
      type: object
      required: [id, motorista_id, placa, capacidade_ton, tipo, is_padrao, created_at]
      properties:
        id:
          type: string
        motorista_id:
          type: string
        placa:
          type: string
        capacidade_ton:
          type: number
        tipo:
          type: string
        is_padrao:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    Espera:
      type: object
      required: [id, motorista_id, veiculo_id, chegada_ts, status, tarifa_ton_hora, created_at]
      properties:
        id:
          type: string
        motorista_id:
          type: string
        veiculo_id:
          type: string
        chegada_ts:
          type: string
          format: date-time
          description: Server-recorded arrival timestamp (UTC)
        chegada_device_ts:
          type: ["string", "null"]
          format: date-time
          description: Device-reported arrival timestamp
        chegada_lat:
          type: ["number", "null"]
        chegada_lng:
          type: ["number", "null"]
        chegada_precisao_m:
          type: ["number", "null"]
        chegada_endereco:
          type: ["string", "null"]
        saida_ts:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
          enum: [aguardando, encerrada, cobranca_gerada]
        embarcador_nome:
          type: ["string", "null"]
        embarcador_cnpj:
          type: ["string", "null"]
        local_descricao:
          type: ["string", "null"]
        fotos:
          type: array
          items:
            type: string
        tarifa_ton_hora:
          type: number
          description: Snapshot of rate at time of record
        valor_calculado:
          type: ["number", "null"]
        veiculo:
          $ref: "#/components/schemas/Veiculo"
        created_at:
          type: string
          format: date-time
    EsperaInput:
      type: object
      required: [veiculo_id, chegada_device_ts, chegada_lat, chegada_lng]
      properties:
        veiculo_id:
          type: string
        chegada_device_ts:
          type: string
          format: date-time
        chegada_lat:
          type: number
        chegada_lng:
          type: number
        chegada_precisao_m:
          type: number
        chegada_endereco:
          type: string
        embarcador_nome:
          type: string
        embarcador_cnpj:
          type: string
        local_descricao:
          type: string
    EncerrarEsperaInput:
      type: object
      required: [saida_ts]
      properties:
        saida_ts:
          type: string
          format: date-time
    EsperaList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Espera"
        total:
          type: integer
...
    FotoInput:
      type: object
      required: [foto_base64, mime_type]
      properties:
        foto_base64:
          type: string
          description: Base64 encoded image data
        mime_type:
          type: string
          description: e.g. image/jpeg
    FotoResult:
      type: object
      required: [url, timestamp]
      properties:
        url:
          type: string
        timestamp:
          type: string
          format: date-time

    # ─── COBRANCAS ──────────────────────────────────────────────────────────
...
    VerificacaoPublica:
      type: object
      required: [cobranca_id, motorista_nome, chegada_ts, local, capacidade_ton, tarifa, valor, registrado_em]
      properties:
        cobranca_id:
          type: string
        motorista_nome:
          type: string
        chegada_ts:
          type: string
          format: date-time
        saida_ts:
          type: ["string", "null"]
          format: date-time
        local:
          type: string
        capacidade_ton:
          type: number
        tarifa:
          type: number
        valor:
          type: number
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        registrado_em:
          type: string
          format: date-time

    # ─── SUBSCRIPTIONS ──────────────────────────────────────────────────────
...