# lib/api-zod/src/generated/types/esperaList.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * ESTADIA API — detention time tracking and billing for truck drivers
 * OpenAPI spec version: 0.1.0
 */
import type { Espera } from './espera';

export interface EsperaList {
  items: Espera[];
  total: number;
}